refactor-app(여행, 관광지): 도메인 디렉토리 분리
