+++ application/lib/domain/place/place_model.dart
