+++ application/lib/domain/place/place_model.freezed.dart
