+++ application/lib/domain/travel/travel_visit_model.freezed.dart
