+++ application/lib/feature/travel_create/provider/travel_create_provider.dart
@@ -1,11 +1,11 @@
 import 'package:application_new/common/http/http_service_provider.dart';
 import 'package:application_new/common/loading/async_loading_provider.dart';
+import 'package:application_new/domain/travel/travel_model.dart';
 import 'package:application_new/feature/authentication/service/auth_service_provider.dart';
 import 'package:application_new/feature/geography/model/city_model.dart';
 import 'package:application_new/feature/geography/model/region_model.dart';
 import 'package:application_new/feature/travel_create/provider/travel_create_state.dart';
-import 'package:application_new/shared/model/travel/travel_form_model.dart';
-import 'package:application_new/shared/model/travel/travel_model.dart';
+import 'package:application_new/domain/travel/travel_model.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'travel_create_provider.g.dart';
