+++ application/lib/feature/travel_plan/city_place_pois/component/city_places_map.dart
@@ -3,7 +3,7 @@ import 'dart:ffi';
 
 import 'package:application_new/feature/travel_plan/city_place_pois/provider/city_place_map_provider.dart';
 import 'package:application_new/feature/travel_read/components/place_marker_item.dart';
-import 'package:application_new/shared/model/place_model.dart';
+import 'package:application_new/domain/place/place_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
