+++ application/lib/feature/travel_plan/city_place_pois/provider/city_place_map_provider.dart
@@ -1,6 +1,6 @@
 import 'package:application_new/feature/travel_plan/city_place_pois/provider/city_place_map_state.dart';
 import 'package:application_new/feature/travel_read/components/place_marker_item.dart';
-import 'package:application_new/shared/model/place_model.dart';
+import 'package:application_new/domain/place/place_model.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
