+++ application/lib/feature/travel_plan/city_travels/page/city_travels_page.dart
@@ -7,7 +7,7 @@ import 'package:application_new/feature/travel_plan/city_travels/provider/city_t
 import 'package:application_new/feature/travel_plan/travel_plan_recommend/component/travel_item.dart';
 import 'package:application_new/shared/component/show_modal_content_sheet.dart';
 import 'package:application_new/shared/component/sliver_infinite_list_indicator.dart';
-import 'package:application_new/shared/model/travel/travel_model.dart';
+import 'package:application_new/domain/travel/travel_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
