+++ application/lib/feature/travel_plan/component/travel_plan_home_header.dart
@@ -1,5 +1,5 @@
 import 'package:application_new/common/util/translation_util.dart';
-import 'package:application_new/shared/model/travel/travel_model.dart';
+import 'package:application_new/domain/travel/travel_model.dart';
 import 'package:flutter/material.dart';
 
 class TravelPlanHomeHeader extends StatelessWidget {
