+++ application/lib/feature/travel_plan/travel_plan_recommend/provider/travel_plan_recommend_state.dart
@@ -1,7 +1,7 @@
 import 'package:application_new/feature/geography/model/city_model.dart';
 import 'package:application_new/feature/travel_plan/travel_plan_recommend/model/place_recommend_model.dart';
-import 'package:application_new/shared/model/place_model.dart';
-import 'package:application_new/shared/model/travel/travel_model.dart';
+import 'package:application_new/domain/place/place_model.dart';
+import 'package:application_new/domain/travel/travel_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'travel_plan_recommend_state.freezed.dart';
