+++ application/lib/feature/travel_read/components/visit_list_item.dart
@@ -1,9 +1,9 @@
 import 'package:application_new/common/util/translation_util.dart';
 import 'package:application_new/feature/travel_read/components/visit_order_item.dart';
-import 'package:application_new/feature/travel_read/model/travel_visit_model.dart';
+import 'package:application_new/domain/travel/travel_visit_model.dart';
 import 'package:application_new/shared/component/small_chip.dart';
-import 'package:application_new/shared/model/place_model.dart';
-import 'package:application_new/shared/provider/place_provider.dart';
+import 'package:application_new/domain/place/place_model.dart';
+import 'package:application_new/domain/place/place_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
