+++ application/lib/feature/travel_read/components/visits_map_item.dart
@@ -1,9 +1,9 @@
 import 'package:application_new/common/loading/loading_page.dart';
 import 'package:application_new/feature/travel_read/components/place_marker_item.dart';
 import 'package:application_new/feature/travel_read/provider/travel_read_provider.dart';
-import 'package:application_new/shared/model/place_model.dart';
-import 'package:application_new/shared/model/travel/travel_model.dart';
-import 'package:application_new/shared/provider/place_provider.dart';
+import 'package:application_new/domain/place/place_model.dart';
+import 'package:application_new/domain/travel/travel_model.dart';
+import 'package:application_new/domain/place/place_provider.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
